Split distance, io and utils into cpp & hpp. Refer ccm.hpp in entrypoint header and update realted refs in sampe & tutorial
